requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0 ; python_version >= '3.7'
attrs==22.2.0 ; python_version >= '3.6'
certifi==2022.12.7 ; python_version >= '3.6'
django==4.1.8
django-debug-toolbar==4.0.0
django-elasticsearch-dsl==7.3
django-extensions==3.2.1
django-filter==23.1
djangorestframework==3.14.0 ; python_version >= '3.6'
djangorestframework-jsonapi==6.0.0
elasticsearch==7.17.9 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3' and python_version < '4'
elasticsearch-dsl==7.4.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
faker==12.0.1 ; python_version >= '3.6'
inflection==0.5.1 ; python_version >= '3.5'
iniconfig==2.0.0 ; python_version >= '3.7'
mixer==7.2.2
packaging==23.0 ; python_version >= '3.7'
pluggy==1.0.0 ; python_version >= '3.6'
pytest==7.2.2
pytest-django==4.5.2
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.3 ; python_version >= '3.5'
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
boto3==1.26.137 
botocore==1.29.137 
jmespath==1.0.1 
s3transfer==0.6.1